opti headers

+++ src/libft/ft_strclr.c
@@ -6,15 +6,13 @@
 /*   By: tdelabro <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 22:29:12 by tdelabro          #+#    #+#             */
-/*   Updated: 2018/11/10 14:01:21 by tdelabro         ###   ########.fr       */
+/*   Updated: 2019/05/25 22:16:34 by tdelabro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
 void	ft_strclr(char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	if (s)
